requirements.txt
flask
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
flask_cors
bcrypt